+++ src/vibeblocks/utils/__init__.py
@@ -1,2 +1,2 @@
-# vibeflow.utils
+# vibeblocks.utils
 # Low-level utilities